LPS-157417 Move logic to new class, no logical changes
